PyLabFastAPI/requirements.txt
aerich==0.9.2
aiofiles==25.1.0
aioredis==2.0.1
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncclick==8.3.0.7
asyncmy==0.2.10
asyncpg==0.31.0
babel==2.17.0
backoff==2.2.1
baidu-aip==4.16.13
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dictdiffer==0.9.0
distro==1.9.0
elastic-transport==9.2.0
elasticsearch==9.2.0
fastapi==0.124.4
fastapi-admin==1.0.4
filelock==3.20.1
fsspec==2025.12.0
googleapis-common-protos==1.72.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib-metadata==8.7.0
iso8601==2.1.0
jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.1
langchain-openai==1.1.3
langfuse==2.60.10
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
markupsafe==3.0.3
numpy==2.3.5
openai==2.12.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.2
passlib==1.7.4
pendulum==3.1.0
pip @ file:///home/task_176313218821961/conda-bld/pip_1763132220821/work
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.10.1
pypika-tortoise==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
qiniu==7.17.0
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scalar-fastapi==1.5.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tortoise-orm==0.25.2
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
uuid-utils==0.12.0
uv==0.9.18
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
